complated homeScrean,category_screan,profile screan and add new file to project
