+++ lib/widgets/banner_slider.dart
@@ -1,3 +1,4 @@
+import "package:apple_shop/constants/colors.dart";
 import "package:flutter/material.dart";
 import "package:smooth_page_indicator/smooth_page_indicator.dart";
 
@@ -7,16 +8,26 @@ class BannerSlider extends StatelessWidget {
   Widget build(BuildContext context) {
     return Stack(alignment: Alignment.bottomCenter, children: [
       SizedBox(
-        height: 200,
+        height: 177,
         child: PageView.builder(
           scrollDirection: Axis.horizontal,
           controller: controller,
           itemBuilder: (context, index) {
             return Container(
               margin: const EdgeInsets.all(8),
-              height: 200,
+              height: 177,
               width: double.infinity,
-              color: Colors.indigoAccent,
+              decoration: BoxDecoration(
+                borderRadius: BorderRadius.circular(15),
+              ),
+              child: const ClipRRect(
+                child: FittedBox(
+                  child: Image(
+                    fit: BoxFit.cover,
+                    image: AssetImage("assets/images/header_banner.png"),
+                  ),
+                ),
+              ),
             );
           },
         ),
@@ -30,7 +41,7 @@ class BannerSlider extends StatelessWidget {
             dotColor: Colors.white,
             dotWidth: 10,
             dotHeight: 10,
-            activeDotColor: Colors.red,
+            activeDotColor: MyColors.blue,
             expansionFactor: 5,
           ), // your preferred effect
         ),
